% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ziehr.2023.ipf.mcti-package.R
\docType{package}
\name{Ziehr.2023.ipf.mcti-package}
\alias{Ziehr.2023.ipf.mcti}
\alias{Ziehr.2023.ipf.mcti-package}
\title{Ziehr.2023.ipf.mcti: Research Compendium}
\description{
Data and analysis code for "Therapeutic potential of lactate transport inhibition in pulmonary fibrosis."
}
\author{
\strong{Maintainer}: William Oldham \email{woldham@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
